requirements/common.txt
celery==4.3.0
Django==2.2.4
setuptools==41.2.0
redis==3.3.8
Pillow==6.1.0
model-mommy==1.6.0
python-decouple==3.1
django-taggit==1.1.0
django-redis==4.10.0
django-celery-results==1.1.2
psycopg2-binary==2.8.3
dj-database-url==0.5.0
django-storages==1.7.1
django-autoslug==1.9.6
gunicorn==19.9.0
boto3==1.9.210